requirements.txt
websockets
google-generativeai
gTTS
miniaudio
